LayoutTests/fast/regions/cssom/offsetParent-in-flow-thread-expected.txt

Test offsetParent for elements inside a named flow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS article.offsetParent is document.body
PASS chapter.offsetParent is document.body
PASS chapter.offsetParent is article
PASS tdChild.offsetParent is td
PASS tdChild.offsetParent is document.body
PASS document.body.offsetParent is null
PASS successfullyParsed is true

TEST COMPLETE
